+++ chabashira_test.go
@@ -66,7 +66,7 @@ type Fragments struct {
 		putMigrate(ts, buf)
 		if buf.String() !=
 		`create_table 'fragments', primary_key:'hidden_pk' do |t|
-  t.integer :id, null:false, limit:8
+  t.bigint :id, null:false, limit:8
   t.integer :version, null:false, default:0, limit:2
   t.integer :size, null:false, limit:4
   t.string :addr, null:false
